+++ browser_tests/setup_and_run_browser_tests.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+set -euf -o pipefail
+
+. setup_browser_tests.sh
+
+. run_browser_tests.sh 
+
