+++ build.xml
@@ -201,7 +201,7 @@ Class-Path: .</echo>
       <arg value="-webcache"/>
       <arg value="false"/>
       <arg value="-max-users"/>
-      <arg value="5"/>
+      <arg value="100"/>
       <arg value="peergos.password"/>
       <arg value="testpassword"/>
       <arg value="pki.keygen.password"/>
