+++ src/components/tour/tour.html
@@ -50,7 +50,7 @@
 		/>
 	      <br/>
 	      <button 
-              id='modal-button-id' 
+              id='skip-tour-modal-button-id' 
 		class="btn btn-danger"
 		style="display: block;float:;"
 		@click="$emit('hide-tour')">
